LPS-52594 move ServiceBuilderArgs out of core
